ports/windows: moved fmode.h and fmode.c to lib/utils/ in order to remove ports/esp32 dependancy on ports/windows.
