+++ lib/utils/fmode.c
