+++ lib/utils/fmode.h
