+++ mpy-cross/Makefile
@@ -65,7 +65,7 @@ SRC_C = \
 # Add fmode when compiling with mingw gcc
 COMPILER_TARGET := $(shell $(CC) -dumpmachine)
 ifneq (,$(findstring mingw,$(COMPILER_TARGET)))
-	SRC_C += ports/windows/fmode.c
+	SRC_C += lib/utils/fmode.c
 endif
 
 OBJ = $(PY_O)
